veryfy user_id when making db queries

+++ internal/storage/postgres/web_writer.go
@@ -23,18 +23,18 @@ func (w *WebWriter) UpsertLesson(ctx context.Context, userID string, lesson *bac
 	var query string
 
 	if lesson.Id > 0 {
-		query = `UPDATE lesson set name = $1, description = $2 where id = $3;`
+		query = `UPDATE lesson SET name = $1, description = $2 WHERE id = $3 AND user_id = $4;`
 
-		_, err := w.db.ExecContext(ctx, query, lesson.Name, lesson.Description, lesson.Id)
+		_, err := w.db.ExecContext(ctx, query, lesson.Name, lesson.Description, lesson.Id, userID)
 		if err != nil {
-			return fmt.Errorf("failed to execute 'UPDATE lesson' query: %w", err)
+			return fmt.Errorf("failed to update lesson: %w", err)
 		}
 	} else {
 		query = `INSERT INTO lesson (name, description, user_id) VALUES ($1, $2, $3) RETURNING id;`
 
 		rows, err := w.db.QueryContext(ctx, query, lesson.Name, lesson.Description, userID)
 		if err != nil {
-			return fmt.Errorf("failed to execute 'INSERT INTO lesson' query: %w", err)
+			return fmt.Errorf("failed to insert lesson: %w", err)
 		}
 		defer rows.Close()
 
@@ -52,11 +52,11 @@ func (w *WebWriter) UpsertLesson(ctx context.Context, userID string, lesson *bac
 func (w *WebWriter) DeleteLesson(ctx context.Context, userID string, lesson backend.Lesson) error {
 	slog.Debug("WebWriter DeleteLesson", "userID", userID)
 
-	query := `DELETE FROM lesson WHERE id = $1;`
+	query := `DELETE FROM lesson WHERE id = $1 AND user_id = $2;`
 
-	_, err := w.db.ExecContext(ctx, query, lesson.Id)
+	_, err := w.db.ExecContext(ctx, query, lesson.Id, userID)
 	if err != nil {
-		return fmt.Errorf("failed to execute 'DELETE FROM lesson' query: %w", err)
+		return fmt.Errorf("failed to delete lesson: %w", err)
 	}
 
 	return nil
@@ -68,18 +68,31 @@ func (w *WebWriter) UpsertExercise(ctx context.Context, userID string, exercise 
 	var query string
 
 	if exercise.Id > 0 {
-		query = `UPDATE exercise set question = $1, answer = $2 where id = $3;`
+		query = `
+			UPDATE exercise e 
+			SET question = $1, answer = $2 
+			FROM lesson l 
+			WHERE e.lesson_id = l.id 
+	  		AND e.id = $3 
+	  		AND l.user_id = $4;
+		`
 
-		_, err := w.db.ExecContext(ctx, query, exercise.Question, exercise.Answer, exercise.Id)
+		_, err := w.db.ExecContext(ctx, query, exercise.Question, exercise.Answer, exercise.Id, userID)
 		if err != nil {
-			return fmt.Errorf("failed to execute 'UPDATE exercise' query: %w", err)
+			return fmt.Errorf("failed to update exercise: %w", err)
 		}
 	} else {
-		query = `INSERT INTO exercise (lesson_id, question, answer) VALUES ($1, $2, $3) RETURNING id;`
+		query = `
+			INSERT INTO exercise (lesson_id, question, answer)
+			SELECT $1, $2, $3
+			FROM lesson l
+			WHERE l.id = $1 AND l.user_id = $4
+			RETURNING id;
+		`
 
-		rows, err := w.db.QueryContext(ctx, query, exercise.Lesson.Id, exercise.Question, exercise.Answer)
+		rows, err := w.db.QueryContext(ctx, query, exercise.Lesson.Id, exercise.Question, exercise.Answer, userID)
 		if err != nil {
-			return fmt.Errorf("failed to execute 'INSERT INTO exercise' query: %w", err)
+			return fmt.Errorf("failed to insert exercise: %w", err)
 		}
 		defer rows.Close()
 
@@ -99,13 +112,16 @@ func (w *WebWriter) StoreExercises(ctx context.Context, userID string, exercises
 
 	const query = `
 		INSERT INTO exercise (lesson_id, question, answer)
-		VALUES ($1, $2, $3)
-		ON CONFLICT (lesson_id, question) DO NOTHING;`
+		SELECT $1, $2, $3
+		FROM lesson l
+		WHERE l.id = $1 AND l.user_id = $4
+		ON CONFLICT (lesson_id, question) DO NOTHING;
+	`
 
 	for _, exercise := range exercises {
-		_, err := w.db.ExecContext(ctx, query, exercise.Lesson.Id, exercise.Question, exercise.Answer)
+		_, err := w.db.ExecContext(ctx, query, exercise.Lesson.Id, exercise.Question, exercise.Answer, userID)
 		if err != nil {
-			return fmt.Errorf("failed to execute 'INSERT INTO exercise' query: %w", err)
+			return fmt.Errorf("failed to store exercise: %w", err)
 		}
 	}
 
@@ -115,11 +131,17 @@ func (w *WebWriter) StoreExercises(ctx context.Context, userID string, exercises
 func (w *WebWriter) DeleteExercise(ctx context.Context, userID string, exercise backend.Exercise) error {
 	slog.Debug("WebWriter DeleteExercise", "userID", userID)
 
-	query := `DELETE FROM exercise WHERE id = $1;`
+	query := `
+		DELETE FROM exercise e
+		USING lesson l
+		WHERE e.lesson_id = l.id
+	  	AND e.id = $1
+	  	AND l.user_id = $2;
+	`
 
-	_, err := w.db.ExecContext(ctx, query, exercise.Id)
+	_, err := w.db.ExecContext(ctx, query, exercise.Id, userID)
 	if err != nil {
-		return fmt.Errorf("failed to execute 'DELETE FROM exercise' query: %w", err)
+		return fmt.Errorf("failed to delete exercise: %w", err)
 	}
 
 	return nil
